Split out VHS fixtures
