+++ pipeline/recorder/src/main/scala/uk/ac/wellcome/platform/recorder/services/RecorderWorkerService.scala
@@ -8,7 +8,7 @@ import io.circe.{Encoder, Json}
 import uk.ac.wellcome.models.work.internal.TransformedBaseWork
 import uk.ac.wellcome.typesafe.Runnable
 
-import uk.ac.wellcome.bigmessaging.typesafe.{EmptyMetadata, GetLocation}
+import uk.ac.wellcome.bigmessaging.{EmptyMetadata, GetLocation}
 import uk.ac.wellcome.messaging.MessageSender
 import uk.ac.wellcome.bigmessaging.message.BigMessageStream
