+++ pipeline/recorder/src/test/scala/uk/ac/wellcome/platform/recorder/fixtures/WorkerServiceFixture.scala
@@ -1,69 +1,27 @@
 package uk.ac.wellcome.platform.recorder.fixtures
 
-import scala.util.{Success, Try}
-
 import uk.ac.wellcome.fixtures.TestWith
 import uk.ac.wellcome.models.Implicits._
 import uk.ac.wellcome.models.work.internal.TransformedBaseWork
 import uk.ac.wellcome.platform.recorder.services.RecorderWorkerService
 
-import uk.ac.wellcome.bigmessaging.typesafe.{EmptyMetadata, GetLocation}
+import uk.ac.wellcome.bigmessaging.EmptyMetadata
 import uk.ac.wellcome.bigmessaging.memory.MemoryTypedStoreCompanion
-import uk.ac.wellcome.bigmessaging.fixtures.BigMessagingFixture
+import uk.ac.wellcome.bigmessaging.fixtures.VHSFixture
 import uk.ac.wellcome.messaging.fixtures.SQS.Queue
-import uk.ac.wellcome.messaging.memory.MemoryMessageSender
 import uk.ac.wellcome.messaging.MessageSender
 
-import uk.ac.wellcome.storage.store.memory.{MemoryStore, MemoryVersionedStore}
 import uk.ac.wellcome.storage.{
   Identified,
   ObjectLocation,
-  StoreWriteError,
   Version
 }
-import uk.ac.wellcome.storage.store.{HybridStoreEntry, VersionedStore}
-import uk.ac.wellcome.storage.maxima.memory.MemoryMaxima
+import uk.ac.wellcome.storage.store.HybridStoreEntry
 
-trait WorkerServiceFixture extends BigMessagingFixture {
-
-  type Entry = HybridStoreEntry[TransformedBaseWork, EmptyMetadata]
-
-  type RecorderVhs = VersionedStore[String, Int, Entry] with GetLocation
-
-  type InternalStore =
-    MemoryStore[Version[String, Int], Entry] with MemoryMaxima[String, Entry]
-
-  class MemoryRecorderVhs(
-    internalStore: InternalStore =
-      new MemoryStore[Version[String, Int], Entry](Map.empty)
-      with MemoryMaxima[String, Entry]
-  ) extends MemoryVersionedStore[String, Entry](internalStore)
-      with GetLocation {
-
-    def getLocation(key: Version[String, Int]): Try[ObjectLocation] =
-      Success(ObjectLocation("test", s"${key.id}/${key.version}"))
-  }
-
-  class BrokenMemoryRecorderVhs extends MemoryRecorderVhs() {
-    override def put(id: Version[String, Int])(entry: Entry): WriteEither =
-      Left(StoreWriteError(new Error("BOOM!")))
-  }
-
-  def withRecorderVhs[R](testWith: TestWith[RecorderVhs, R]): R = {
-    testWith(new MemoryRecorderVhs())
-  }
-
-  def withBrokenRecorderVhs[R](testWith: TestWith[RecorderVhs, R]): R = {
-    testWith(new BrokenMemoryRecorderVhs())
-  }
-
-  def withMemoryMessageSender[R](
-    testWith: TestWith[MemoryMessageSender, R]): R = {
-    testWith(new MemoryMessageSender())
-  }
+trait WorkerServiceFixture extends VHSFixture[TransformedBaseWork] {
 
   def withWorkerService[R, D](queue: Queue,
-                              vhs: RecorderVhs,
+                              vhs: VHS,
                               msgSender: MessageSender[D])(
     testWith: TestWith[RecorderWorkerService[D], R]): R =
     withActorSystem { implicit actorSystem =>
@@ -77,7 +35,7 @@ trait WorkerServiceFixture extends BigMessagingFixture {
     }
 
   def assertWorkStored[T <: TransformedBaseWork](
-    vhs: RecorderVhs,
+    vhs: VHS,
     work: T,
     expectedVhsVersion: Int = 0): Version[String, Int] = {
 
@@ -90,7 +48,7 @@ trait WorkerServiceFixture extends BigMessagingFixture {
     Version(id, expectedVhsVersion)
   }
 
-  def assertWorkNotStored[T <: TransformedBaseWork](vhs: RecorderVhs,
+  def assertWorkNotStored[T <: TransformedBaseWork](vhs: VHS,
                                                     work: T) = {
 
     val id = work.sourceIdentifier.toString
